requirements.txt
cycler==0.10.0
html2text==2016.9.19
lxml==3.7.3
matplotlib==2.0.2
numpy==1.12.1
olefile==0.44
Pillow==4.1.1
pyparsing==2.2.0
python-dateutil==2.6.0
python-pptx==0.6.5
pytz==2017.2
six==1.10.0
wheel==0.24.0
wordcloud==1.3.1
XlsxWriter==0.9.6
